Update benchmark scripts

- Update the brian2genn and genn submodules, which were set to a wrong commit
- Add scrips to easily run benchmarks on our cluster
- Fix benchmark script issues on the way

+++ brian2cuda/tools/remote_run_scripts/_on_headnode.sh
@@ -1,29 +1,48 @@
 #!/bin/bash
-# $1:    path to brian2cuda git repository which should be run as
-# $2:    entire path to logfile
-# $3:    path to conda.sh
-# $4:    conda env name
-# $5...: the rest is passed as args to run_test_suite.py
-b2c_dir="$1"
-logfile="$2"
-path_conda_sh="$3"
-conda_env="$4"
-shift 4
+# Don't call this file directly, it is used for submitting test suite and
+# benchmark scripts to the grid engine (see e.g. run_test_suite_on_cluster.sh)
+# It will set the conda environment, initialize CUDA variables and make sure
+# the previously copied brian2cuda repository is deleted after the run.
+
+# $1:    path to bash script that will be submitted to grid engine, relative to
+#        brian2cuda repository path ($2).
+#        Example: brian2cuda/tools/test_suite/_run_test_suite.sh
+# $2:    path to brian2cuda git repository which should be run as
+# $3:    entire path to logfile
+# $4:    path to conda.sh
+# $5:    conda env name
+# $6...: the rest is passed as args to run_test_suite.py
+bash_script="$1"
+b2c_dir="$2"
+logfile="$3"
+path_conda_sh="$4"
+conda_env="$5"
+keep_remote_repo="$6"
+shift 6
 test_suite_args="$@"
 
 # deletes the brian2cuda directory
 function cleanup {
     cd
-    rm -rf "$b2c_dir"
-    echo "Script EXIT: Deleted tmp brian2cuda directory $b2c_dir" | tee -a "$logfile"
+    if [ $keep_remote_repo -eq 1 ]; then
+        rm -rf "$b2c_dir"
+        echo "Script EXIT: Deleted tmp brian2cuda directory $b2c_dir" | tee -a "$logfile"
+    else
+        echo "Script EXIT: tmp brian2cuda directory was not deleted" \
+             "(-k|--keep-remote-dir option). If you don't need it, delete it:" \
+             "$b2c_dir" | tee -a "$logfile"
+    fi
 }
 
 # register the cleanup function to be called on the EXIT signal
 trap cleanup EXIT
 
-# where to save log files
-logdir="$(dirname $logfile)"
-run_name="$(basename $logfile)"
+# Get directory and name of logfile. For test suite script, the logfile has the
+# timestemp and the directory is where all different logfiles are stored. For
+# benchmark scripts, the directory has the timestemp and the logfile is always
+# `full.log`
+logfile_dir="$(dirname $logfile)"
+logfile_name="$(basename $logfile)"
 
 # activate bashrc (for conda activation and CUDA paths)
 . "$path_conda_sh"
@@ -36,6 +55,9 @@ else
     . /cognition/home/local/.init_cuda.sh
 fi
 
+script_path="$b2c_dir/$bash_script"
+script_folder="$(dirname $script_path)"
+script_name="$(basename $script_path)"
 # XXX: needs to cd into the tools directory for PYTHONPATH setup to work
-cd "$b2c_dir"/brian2cuda/tools
-bash _run_test_suite.sh "$run_name" "$logdir" "$test_suite_args"
+cd "$script_folder"
+bash "$script_name" "$logfile_name" "$logfile_dir" "$test_suite_args"
